refactor(switches): rename package name from `option` to `switches`
