+++ app/src/main/java/com/osfans/trime/ime/bar/QuickBar.kt
@@ -30,7 +30,7 @@ import com.osfans.trime.ime.keyboard.CommonKeyboardActionListener
 import com.osfans.trime.ime.keyboard.GestureFrame
 import com.osfans.trime.ime.keyboard.InputFeedbackManager
 import com.osfans.trime.ime.keyboard.KeyboardWindow
-import com.osfans.trime.ime.option.SwitchOptionWindow
+import com.osfans.trime.ime.switches.SwitchOptionWindow
 import com.osfans.trime.ime.window.BoardWindow
 import com.osfans.trime.ime.window.BoardWindowManager
 import me.tatarka.inject.annotations.Inject
