+++ app/src/main/java/com/osfans/trime/ime/switches/SwitchOptionWindow.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: GPL-3.0-or-later
  */
 
-package com.osfans.trime.ime.option
+package com.osfans.trime.ime.switches
 
 import android.content.Context
 import android.view.View
